+++ src/normalizer.ts
@@ -1,9 +1,18 @@
 import {JSONSchemaTypeName, LinkedJSONSchema, NormalizedJSONSchema, Parent} from './types/JSONSchema'
-import {escapeBlockComment, justName, toSafeString, traverse} from './utils'
+import {appendToDescription, escapeBlockComment, isSchemaLike, justName, toSafeString, traverse} from './utils'
 import {Options} from './'
 import {isPlainObject} from 'lodash'
+import {applySchemaTyping} from './applySchemaTyping'
+import {DereferencedPaths} from './resolver'
+import {isDeepStrictEqual} from 'util'
 
-type Rule = (schema: LinkedJSONSchema, fileName: string, options: Options) => void
+type Rule = (
+  schema: LinkedJSONSchema,
+  fileName: string,
+  options: Options,
+  key: string | null,
+  dereferencedPaths: DereferencedPaths,
+) => void
 const rules = new Map<string, Rule>()
 
 function hasType(schema: LinkedJSONSchema, type: JSONSchemaTypeName) {
@@ -14,6 +23,9 @@ function isObjectType(schema: LinkedJSONSchema) {
 }
 function isArrayType(schema: LinkedJSONSchema) {
   return schema.items !== undefined || hasType(schema, 'array') || hasType(schema, 'any')
+}
+function isEnumTypeWithoutTsEnumNames(schema: LinkedJSONSchema) {
+  return schema.type === 'string' && schema.enum !== undefined && schema.tsEnumNames === undefined
 }
 
 rules.set('Remove `type=["null"]` if `enum=[null]`', schema => {
@@ -45,32 +57,81 @@ rules.set('Transform `required`=false to `required`=[]', schema => {
   }
 })
 
-// TODO: default to empty schema (as per spec) instead
-rules.set('Default additionalProperties to true', schema => {
+rules.set('Default additionalProperties', (schema, _, options) => {
   if (isObjectType(schema) && !('additionalProperties' in schema) && schema.patternProperties === undefined) {
-    schema.additionalProperties = true
+    schema.additionalProperties = options.additionalProperties
   }
 })
 
-rules.set('Default top level `id`', (schema, fileName) => {
-  const isRoot = schema[Parent] === null
-  if (isRoot && !schema.id) {
-    schema.id = toSafeString(justName(fileName))
+rules.set('Transform id to $id', (schema, fileName) => {
+  if (!isSchemaLike(schema)) {
+    return
+  }
+  if (schema.id && schema.$id && schema.id !== schema.$id) {
+    throw ReferenceError(
+      `Schema must define either id or $id, not both. Given id=${schema.id}, $id=${schema.$id} in ${fileName}`,
+    )
+  }
+  if (schema.id) {
+    schema.$id = schema.id
+    delete schema.id
   }
 })
 
-rules.set('Escape closing JSDoc Comment', schema => {
+rules.set('Add an $id to anything that needs it', (schema, fileName, _options, _key, dereferencedPaths) => {
+  if (!isSchemaLike(schema)) {
+    return
+  }
+
+  // Top-level schema
+  if (!schema.$id && !schema[Parent]) {
+    schema.$id = toSafeString(justName(fileName))
+    return
+  }
+
+  // Sub-schemas with references
+  if (!isArrayType(schema) && !isObjectType(schema)) {
+    return
+  }
+
+  // We'll infer from $id and title downstream
+  // TODO: Normalize upstream
+  const dereferencedName = dereferencedPaths.get(schema)
+  if (!schema.$id && !schema.title && dereferencedName) {
+    schema.$id = toSafeString(justName(dereferencedName))
+  }
+
+  if (dereferencedName) {
+    dereferencedPaths.delete(schema)
+  }
+})
+
+rules.set('Escape closing JSDoc comment', schema => {
   escapeBlockComment(schema)
 })
 
+rules.set('Add JSDoc comments for minItems and maxItems', schema => {
+  if (!isArrayType(schema)) {
+    return
+  }
+  const commentsToAppend = [
+    'minItems' in schema ? `@minItems ${schema.minItems}` : '',
+    'maxItems' in schema ? `@maxItems ${schema.maxItems}` : '',
+  ].filter(Boolean)
+  if (commentsToAppend.length) {
+    schema.description = appendToDescription(schema.description, ...commentsToAppend)
+  }
+})
+
 rules.set('Optionally remove maxItems and minItems', (schema, _fileName, options) => {
-  if (options.ignoreMinAndMaxItems) {
-    if (isPlainObject(schema) && 'maxItems' in schema) {
-      delete schema.maxItems
-    }
-    if (isPlainObject(schema) && 'minItems' in schema) {
-      delete schema.minItems
-    }
+  if (!isArrayType(schema)) {
+    return
+  }
+  if (isPlainObject(schema) && 'minItems' in schema && options.ignoreMinAndMaxItems) {
+    delete schema.minItems
+  }
+  if (isPlainObject(schema) && 'maxItems' in schema && (options.ignoreMinAndMaxItems || options.maxItems === -1)) {
+    delete schema.maxItems
   }
 })
 
@@ -79,11 +140,26 @@ rules.set('Normalize schema.minItems', (schema, _fileName, options) => {
     return
   }
   // make sure we only add the props onto array types
-  if (isArrayType(schema)) {
-    const {minItems} = schema
-    schema.minItems = typeof minItems === 'number' ? minItems : 0
+  if (!isArrayType(schema)) {
+    return
   }
+  const {minItems} = schema
+  schema.minItems = typeof minItems === 'number' ? minItems : 0
   // cannot normalize maxItems because maxItems = 0 has an actual meaning
+})
+
+rules.set('Remove maxItems if it is big enough to likely cause OOMs', (schema, _fileName, options) => {
+  if (options.ignoreMinAndMaxItems || options.maxItems === -1) {
+    return
+  }
+  if (!isArrayType(schema)) {
+    return
+  }
+  const {maxItems, minItems} = schema
+  // minItems is guaranteed to be a number after the previous rule runs
+  if (maxItems !== undefined && maxItems - (minItems as number) > options.maxItems) {
+    delete schema.maxItems
+  }
 })
 
 rules.set('Normalize schema.items', (schema, _fileName, options) => {
@@ -132,10 +208,15 @@ rules.set('Make extends always an array, if it is defined', schema => {
   }
 })
 
-rules.set('Transform $defs to definitions', schema => {
-  if (schema.$defs) {
-    schema.definitions = schema.$defs
-    delete schema.$defs
+rules.set('Transform definitions to $defs', (schema, fileName) => {
+  if (schema.definitions && schema.$defs && !isDeepStrictEqual(schema.definitions, schema.$defs)) {
+    throw ReferenceError(
+      `Schema must define either definitions or $defs, not both. Given id=${schema.id} in ${fileName}`,
+    )
+  }
+  if (schema.definitions) {
+    schema.$defs = schema.definitions
+    delete schema.definitions
   }
 })
 
@@ -146,7 +227,34 @@ rules.set('Transform const to singleton enum', schema => {
   }
 })
 
-export function normalize(rootSchema: LinkedJSONSchema, filename: string, options: Options): NormalizedJSONSchema {
-  rules.forEach(rule => traverse(rootSchema, schema => rule(schema, filename, options)))
+rules.set('Add tsEnumNames to enum types', (schema, _, options) => {
+  if (isEnumTypeWithoutTsEnumNames(schema) && options.inferStringEnumKeysFromValues) {
+    schema.tsEnumNames = schema.enum?.map(String)
+  }
+})
+
+// Precalculation of the schema types is necessary because the ALL_OF type
+// is implemented in a way that mutates the schema object. Detection of the
+// NAMED_SCHEMA type relies on the presence of the $id property, which is
+// hoisted to a parent schema object during the ALL_OF type implementation,
+// and becomes unavailable if the same schema is used in multiple places.
+//
+// Precalculation of the `ALL_OF` intersection schema is necessary because
+// the intersection schema needs to participate in the schema cache during
+// the parsing step, so it cannot be re-calculated every time the schema
+// is encountered.
+rules.set('Pre-calculate schema types and intersections', schema => {
+  if (schema !== null && typeof schema === 'object') {
+    applySchemaTyping(schema)
+  }
+})
+
+export function normalize(
+  rootSchema: LinkedJSONSchema,
+  dereferencedPaths: DereferencedPaths,
+  filename: string,
+  options: Options,
+): NormalizedJSONSchema {
+  rules.forEach(rule => traverse(rootSchema, (schema, key) => rule(schema, filename, options, key, dereferencedPaths)))
   return rootSchema as NormalizedJSONSchema
 }