+++ test/testE2E.ts
@@ -1,9 +1,11 @@
+import type {FileInfo} from '@apidevtools/json-schema-ref-parser'
 import test from 'ava'
 import {readdirSync} from 'fs'
-import {find} from 'lodash'
+import {find, merge} from 'lodash'
 import {join} from 'path'
 import {compile, JSONSchema, Options} from '../src'
 import {log, stripExtension} from '../src/utils'
+import {getWithCache} from './http'
 
 const dir = __dirname + '/e2e'
 
@@ -39,19 +41,30 @@ export function run() {
   }
 }
 
+const httpWithCacheResolver = {
+  order: 1,
+  canRead: /^https?:/i,
+  async read({url}: FileInfo) {
+    return await getWithCache(url)
+  },
+}
+
 function runOne(exports: TestCase, name: string) {
   log('blue', 'Running test', name)
+
+  const options = merge(exports.options, {$refOptions: {resolve: {http: httpWithCacheResolver}}})
+
   test(name, async t => {
     if (exports.error) {
       try {
-        await compile(exports.input, stripExtension(name), exports.options)
+        await compile(exports.input, stripExtension(name), options)
       } catch (e) {
         t.true(e instanceof Error)
       }
     } else {
       t.snapshot(
-        await compile(exports.input, stripExtension(name), exports.options),
-        `Expected output to match snapshot for e2e test: ${name}`
+        await compile(exports.input, stripExtension(name), options),
+        `Expected output to match snapshot for e2e test: ${name}`,
       )
     }
   })